refactor: move prompt utils into a `utils` module
